Move shared file walker to kernel fs-walk module
